change handling of the CODE items for Finding/Finding site
In the past, rows corresponding to those concepts were excluded,
since they are included as columns. With the introduction of the
nnU-Net-BPR-annotations collection, we have SRs that contain "Finding
site" as the kind of slice-level annotations. To make the query
consistent, I decided to now include all CODE items included in the
measurement group into the qualitative assessments table.